+++ spring-integration/spring-boot-autoconfigure/src/test/java/ai/timefold/solver/spring/boot/autoconfigure/TimefoldSolverGizmoAutoConfigurationTest.java
@@ -5,9 +5,11 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import ai.timefold.solver.core.api.domain.common.DomainAccessType;
 import ai.timefold.solver.core.api.solver.SolverFactory;
+import ai.timefold.solver.core.api.solver.SolverManager;
 import ai.timefold.solver.core.config.solver.SolverConfig;
 import ai.timefold.solver.spring.boot.autoconfigure.config.TimefoldProperties;
 import ai.timefold.solver.spring.boot.autoconfigure.gizmo.GizmoSpringTestConfiguration;
+import ai.timefold.solver.spring.boot.autoconfigure.normal.domain.TestdataSpringSolution;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
@@ -43,6 +45,17 @@ class TimefoldSolverGizmoAutoConfigurationTest {
                     var solverConfig = context.getBean(SolverConfig.class);
                     assertThat(solverConfig.getDomainAccessType()).isEqualTo(DomainAccessType.GIZMO);
                     assertThat(context.getBean(SolverFactory.class)).isNotNull();
+                });
+        gizmoContextRunner
+                .withPropertyValues("timefold.solver.solver1.domain-access-type=GIZMO")
+                .withPropertyValues("timefold.solver.solver2.domain-access-type=REFLECTION")
+                .run(context -> {
+                    var solver1 =
+                            (SolverManager<TestdataSpringSolution, Long>) context.getBean("solver1");
+                    var solver2 =
+                            (SolverManager<TestdataSpringSolution, Long>) context.getBean("solver2");
+                    assertThat(solver1).isNotNull();
+                    assertThat(solver2).isNotNull();
                 });
     }
